fixed loss of precision warning so it compiles

+++ src/components.cpp
@@ -75,15 +75,15 @@ bool Mesh::loadFromOBJFile(std::string obj_path, std::vector<ColoredVertex>& out
 //                    out_normal_indices.push_back((uint16_t)(normalIndex[i] - 1));
 //                }
                 // -1 since .obj starts counting at 1 and OpenGL starts at 0
-                out_vertex_indices.push_back((uint16_t)vertexIndex[0] - 1);
-                out_vertex_indices.push_back((uint16_t)vertexIndex[1] - 1);
-                out_vertex_indices.push_back((uint16_t)vertexIndex[2] - 1);
-                out_uv_indices.push_back(uvIndex[0] - 1);
-                out_uv_indices.push_back(uvIndex[1] - 1);
-                out_uv_indices.push_back(uvIndex[2] - 1);
-                out_normal_indices.push_back((uint16_t)normalIndex[0] - 1);
-                out_normal_indices.push_back((uint16_t)normalIndex[1] - 1);
-                out_normal_indices.push_back((uint16_t)normalIndex[2] - 1);
+                out_vertex_indices.push_back(static_cast<uint16_t>(vertexIndex[0]) - 1);
+                out_vertex_indices.push_back(static_cast<uint16_t>(vertexIndex[1]) - 1);
+                out_vertex_indices.push_back(static_cast<uint16_t>(vertexIndex[2]) - 1);
+                out_uv_indices.push_back(static_cast<uint16_t>(uvIndex[0]) - 1);
+                out_uv_indices.push_back(static_cast<uint16_t>(uvIndex[1]) - 1);
+                out_uv_indices.push_back(static_cast<uint16_t>(uvIndex[2]) - 1);
+                out_normal_indices.push_back(static_cast<uint16_t>(normalIndex[0]) - 1);
+                out_normal_indices.push_back(static_cast<uint16_t>(normalIndex[1]) - 1);
+                out_normal_indices.push_back(static_cast<uint16_t>(normalIndex[2]) - 1);
             }
 //            else {
 //                // Attempt to read `v//n` format
